refactor: Name middleware
