+++ src/routers/loginRouter.ts
@@ -1,10 +1,10 @@
 import { Router } from "express"
 
 import { loginController } from "../modules/login/controllers"
-import { validationsLogin } from "../modules/login/middlewares"
+import { validationLogin } from "../modules/login/middlewares"
 
 const loginRouter = Router()
 
-loginRouter.post('/', validationsLogin.validationFields, loginController.login)
+loginRouter.post('/', validationLogin.validationFields, loginController.login)
 
 export { loginRouter }
